% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_duplicate}
\alias{create_duplicate}
\title{Duplicate an R object}
\usage{
create_duplicate(x)
}
\arguments{
\item{x}{An R object to duplicate. Must not be \code{NULL}.}
}
\value{
A duplicate of \code{x}.
}
\description{
Creates a duplicate of an R object using R's native duplication mechanism.
}
