% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cure.R
\name{cure}
\alias{cure}
\title{Recover an infected function inside a namespace}
\usage{
cure(x, ...)
}
\arguments{
\item{x}{character or a function object}

\item{ns}{where to search for x}
}
\value{
the infected x
}
\description{
Recover an infected function inside a namespace
}
