% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{.object_wrapper}
\alias{.object_wrapper}
\title{Wrapper for object to allow storing of arbitrary configuration information}
\usage{
.object_wrapper(object, ...)
}
\description{
Wrapper for object to allow storing of arbitrary configuration information
}
