% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eject.R
\name{eject}
\alias{eject}
\title{Eject function source to a file for manipulation}
\usage{
eject(x, from = parent.frame(), to, ...)
}
\arguments{
\item{x}{a string of function name or an object of class function}

\item{from}{where to search for x}

\item{to}{file path to dump x}

\item{...}{ignore}
}
\value{
invisible the name of x
}
\description{
Simple function that eject function from a namespace to a file, calling dump
}
