% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inject.R
\name{inject}
\alias{inject}
\title{Inject a function into a package namespace}
\usage{
inject(
  x,
  name = deparse(substitutex(x)),
  target,
  infect = TRUE,
  curable = FALSE,
  ...
)
}
\arguments{
\item{x}{the function value, can be a file}

\item{name}{the name of the inject object}

\item{target}{<character(1)> the target namespace so that the function will operate on}

\item{infect}{whether to assign itself into that namespace. if FALSE, infect is similar to `environment(x) <- asNamespace(target)`.
After the infection, users can call x by `target:::name`. Very effective when you are modifying a method.}

\item{curable}{whether to keep a copy of the original function if x is overwriting an existing method}
}
\value{
x, invisibly if infect=TRUE
}
\description{
Inject a function into a package namespace
}
