% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{register}
\alias{register}
\title{Register an object to inject}
\usage{
register(object, name = NULL)
}
\arguments{
\item{object}{the object to inject}

\item{name}{the name to use for resolution. Defaults to the passed in object}
}
\description{
Register an object to inject
}
