% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sexp_address}
\alias{sexp_address}
\title{Get the memory address of an R object}
\usage{
sexp_address(x)
}
\arguments{
\item{x}{An R object. Must not be \code{NULL}.}
}
\value{
A character scalar containing the hexadecimal memory address of \code{x}.
}
\description{
Returns the address of an object's underlying S-expression. This low-level
helper is primarily useful when testing whether two objects share storage.
}
