% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{inject}
\alias{inject}
\title{Inject an object from the container into a variable}
\usage{
inject(name, target = NULL)
}
\arguments{
\item{name}{the name of the object to inject}
}
\value{
the defined object or null
}
\description{
Inject an object from the container into a variable
}
